misc(incoming_webhook): Refact handling for Adyen and Cashfree
