+++ app/models/payment_providers/adyen_provider.rb
@@ -4,6 +4,8 @@ module PaymentProviders
   class AdyenProvider < BaseProvider
     SUCCESS_REDIRECT_URL = "https://www.adyen.com/"
 
+    WEBHOOKS_EVENTS = %w[AUTHORISATION REFUND REFUND_FAILED CHARGEBACK].freeze
+
     PROCESSING_STATUSES = %w[AuthorisedPending Received].freeze
     SUCCESS_STATUSES = %w[Authorised SentForSettle SettleScheduled Settled Refunded].freeze
     FAILED_STATUSES = %w[Cancelled CaptureFailed Error Expired Refused].freeze
