+++ spec/services/payment_providers/cashfree_service_spec.rb
@@ -107,46 +107,4 @@ RSpec.describe PaymentProviders::CashfreeService, type: :service do
       end
     end
   end
-
-  describe ".handle_incoming_webhook" do
-    let(:cashfree_provider) { create(:cashfree_provider, organization:, client_id:, client_secret:) }
-
-    let(:body) do
-      path = Rails.root.join("spec/fixtures/cashfree/payment_link_event_payment.json")
-      JSON.parse(File.read(path)).to_json # NOTE: Ensure valid sha256 signature
-    end
-
-    before { cashfree_provider }
-
-    it "checks the webhook" do
-      result = cashfree_service.handle_incoming_webhook(
-        organization_id: organization.id,
-        body:,
-        timestamp: "1629271506",
-        signature: "MFB3Rkubs4jB97ROS/I4iu9llAAP5ykJ3GZYp95o/Mw="
-      )
-
-      expect(result).to be_success
-
-      expect(PaymentProviders::Cashfree::HandleEventJob).to have_been_enqueued
-    end
-
-    context "when failing to validate the signature" do
-      it "returns an error" do
-        result = cashfree_service.handle_incoming_webhook(
-          organization_id: organization.id,
-          body:,
-          timestamp: "1629271506",
-          signature: "signature"
-        )
-
-        aggregate_failures do
-          expect(result).not_to be_success
-          expect(result.error).to be_a(BaseService::ServiceFailure)
-          expect(result.error.code).to eq("webhook_error")
-          expect(result.error.error_message).to eq("Invalid signature")
-        end
-      end
-    end
-  end
 end
